fix more type issues
