+++ app/models/colyseus-models/synergies.ts
@@ -1,15 +1,13 @@
 import { MapSchema, SetSchema } from "@colyseus/schema"
 import { SynergyTriggers } from "../../config"
-import { type IPlayer, type IPokemon, Item } from "../../types"
+import { Item } from "../../types"
 import { SynergyGivenByItem } from "../../types/enum/Item"
 import { Passive } from "../../types/enum/Passive"
 import { Pkm, PkmFamily, PkmIndex } from "../../types/enum/Pokemon"
 import { SpecialGameRule } from "../../types/enum/SpecialGameRule"
 import { Synergy, SynergyArray } from "../../types/enum/Synergy"
-import { isOnBench } from "../../utils/board"
 import { schemaValues } from "../../utils/schemas"
 import { getPokemonData } from "../precomputed/precomputed-pokemon-data"
-import { PVEStages } from "../pve-stages"
 import type { Pokemon } from "./pokemon"
 
 export default class Synergies extends MapSchema<number, Synergy> {
@@ -84,7 +82,7 @@ export function computeSynergies(
 
   const typesPerFamily = new Map<string, Set<Synergy>>()
 
-  board.forEach((pkm: IPokemon, index) => {
+  board.forEach((pkm: Pokemon, index) => {
     // reset dynamic synergies
     if (pkm.passive === Passive.PROTEAN2 || pkm.passive === Passive.PROTEAN3) {
       pkm.types.clear()
@@ -289,7 +287,7 @@ export function computeSynergies(
   return synergies
 }
 
-export function addSynergiesGivenByItems(pkm: IPokemon) {
+export function addSynergiesGivenByItems(pkm: Pokemon) {
   pkm.items.forEach((item) => {
     const synergy = SynergyGivenByItem[item]
     if (synergy) {
@@ -314,25 +312,4 @@ export function addSynergiesGivenByItems(pkm: IPokemon) {
       }
     }
   })
-}
-
-export function getSynergyStep(
-  synergies: Map<Synergy, number> | MapSchema<number, Synergy>,
-  type: Synergy
-): number {
-  return SynergyTriggers[type].filter((n) => (synergies.get(type) ?? 0) >= n)
-    .length
-}
-
-export function getWildChance(player: IPlayer, stageLevel: number): number {
-  const isPVE = stageLevel === 0 || stageLevel in PVEStages
-  const wildLevel = getSynergyStep(player.synergies, Synergy.WILD)
-  // 6% base chance in PvE stage or if Wild is active
-  const baseChance = isPVE || wildLevel > 0 ? 6 : 0
-  // each star of a pokemon with wild synergy gives 0.5% wild chance
-  const nbWildStars = schemaValues(player.board)
-    .filter((p) => p.hasSynergy(Synergy.WILD) && isOnBench(p) === false)
-    .reduce((total, p) => total + p.stars, 0)
-  const bonusChance = wildLevel > 0 ? nbWildStars * 0.5 : 0
-  return (baseChance + bonusChance) / 100
 }
