+++ app/models/shop.ts
@@ -41,6 +41,7 @@ import {
   UniquePool
 } from "../config"
 import { pickFirstPartners } from "../core/scribbles"
+import { getWildChance } from "../core/synergies"
 import type GameState from "../rooms/states/game-state"
 import type { IPokemon, IPokemonEntity } from "../types"
 import { EffectEnum } from "../types/enum/Effect"
@@ -78,7 +79,6 @@ import {
   type PlayerChoiceType
 } from "./colyseus-models/player-choice"
 import { type Pokemon, PokemonClasses } from "./colyseus-models/pokemon"
-import { getWildChance } from "./colyseus-models/synergies"
 import { getPokemonBaseline } from "./pokemon-factory"
 import {
   getPokemonData,
