+++ app/public/src/pages/component/synergy/synergy-detail-component.tsx
@@ -4,7 +4,7 @@ import {
   type SynergyEffect,
   SynergyEffects
 } from "../../../../../config/game/synergies"
-import { getWildChance } from "../../../../../models/colyseus-models/synergies"
+import { getWildChance } from "../../../../../core/synergies"
 import { getPokemonData } from "../../../../../models/precomputed/precomputed-pokemon-data"
 import { PRECOMPUTED_POKEMONS_PER_TYPE_AND_CATEGORY } from "../../../../../models/precomputed/precomputed-types-and-categories"
 import type { IPlayer } from "../../../../../types"
@@ -111,7 +111,9 @@ export default function SynergyDetailComponent(props: {
         const totalDragonStars = schemaValues(spectatedPlayer.board).reduce(
           (acc, pokemon) =>
             acc +
-            (pokemon.hasSynergy(Synergy.DRAGON) && !isOnBench(pokemon)
+            ((pokemon.types.has(Synergy.DRAGON) ||
+              pokemon.types.has(Synergy.STELLAR)) &&
+            !isOnBench(pokemon)
               ? pokemon.stars
               : 0),
           0
@@ -227,8 +229,8 @@ function PokemonPortraitList(props: {
     props.player == null
       ? []
       : schemaValues(props.player.board)
-          .filter((x) => x.hasSynergy(props.type))
-          .map((x) => PkmFamily[x.name])
+          .filter((p) => p.types.has(props.type))
+          .map((p) => PkmFamily[p.name])
   )
 
   return (
